Take ownership of other value
Previosuly, the merge method of the Merge trait took a mutable reference
to the other value.  This patch changes the method to take ownership of
the value instead.  Firstly, this makes it clear that we merge other
into self and not self into other.  Secondly, this allows us to avoid
copies when moving the other value into self.