Remove unnecessary trait requirements
Previously, the right argument for a merge strategy was a reference, so
we had to make sure that it implements copy so that we can assign it to
left.  But as we now take ownership of the argument, we no longer need
the Copy requirement for the generic merge strategies.